+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * Replace logrus with zerolog. (#413, @miry)
 * Log HTTP requests to API server. (#413, #421, @miry)
 * Add TimeoutHandler for the HTTP API server. (#420, @miry)
-* Set Write and Read timeouts for HTTP API server connections. (@miry)
+* Set Write and Read timeouts for HTTP API server connections. (#423, @miry)
+* Show uniq request id in API HTTP response. (#425, @miry)
 
 # [2.4.0] - 2022-03-07